refactor: add catalog component
