+++ src/components/catalog/index.ts
@@ -1,0 +1,1 @@
+export { default as Catalog } from './catalog';
